backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
mistralai>=0.1.0